package modularization wrt argparse
