--- .gitignore
@@ -1,2 +1,0 @@
-# module file
-go.mod
