+++ go.mod
@@ -1,0 +1,3 @@
+module example.com/m
+
+go 1.22.5
