+++ src/fl.go
@@ -7,117 +7,17 @@ import (
 	"os"
 	"os/exec"
 	"strings"
-	"errors"
+	"fl/helpers"
 )
 
 /**********************
- * globals/helpers
+ * globals
  *********************/
-
-// global flag structure
-type FlagStruct struct {
-	verbose, help, autoexecute, noexec bool
-	len int
-}
 
 const (
 	// url of the Flow endpoint
 	apiUrl = "https://flow.pstmn-beta.io/api/4e5b4cfcdec54831a31d9f38aaf1a938"
 ) 
-
-func new_flags()(Flags FlagStruct) {
-	return FlagStruct {
-		verbose: false,
-		help: false,
-		autoexecute: false,
-		noexec: false,
-		len: 4,
-	}
-}
-
-/*
- * To add a new flag:
- * 1. Update its entry in Usage()
- * 2. Add it to the list of flags above and update numFlags
- * 3. Create a handler function and add it to the switch-case in argParse()
- */
-// useage definition functions to explain command and its args
-var Usage = func () {
-	fmt.Println("Usage: fl [-hynv] prompt...")
-	// for formatting - please start with a space and ensure descruption alignment with tabs
-	fmt.Println(" -h,--help\t\tshow command usage")
-	fmt.Println(" -y\t\t\tautoexecute the generated command")
-	fmt.Println(" -n\t\t\tdo not prompt for or run generated command (takes priority over -y)")
-	fmt.Println(" -v,--verbose\t\tdisplay updates of the command progress")
-}
-
-// print passed prompt if verbose check set
-func print(verbose bool, str ...interface{}) {
-	if verbose {
-		fmt.Println(str...)
-	}
-}
-
-/**********************
- * arg parsing
- *********************/
-
-// parse the user input for potential prompts
-var argParse = func (args []string, Flags *FlagStruct) (prompt string, err error) {
-	// Check if command line arguments are provided
-	if len(args) < 2 {
-		// expecting at least 2 arguments
-		return "", errors.New("expecting at least 2 args")
-	}
-
-	// Start of the user prompt (after args have been parsed)
-	startPromptIndex := 1
-	// flag to exit for loop if non-flag detected
-	validArg := true
-
-	/* check for flags (add 1 bc first index is command path)
-	 * validarg: if non-flag found, check for prompt
-	 * also exit if args runs out
-	 */
-	for i := 1; i < Flags.len+1 && i < len(args) && validArg; i++ {
-		switch args[i] {
-		case "-h": // help commands (just display useage)
-			fallthrough
-		case "--help":
-			startPromptIndex++
-			Flags.help = true
-			return "", nil // exit early if help flag
-		case "-v": // handle program verbosity
-			fallthrough
-		case "--verbose":
-			startPromptIndex++
-			Flags.verbose = true
-		case "-y": // execute command automatically
-			startPromptIndex++
-			Flags.autoexecute = true
-		case "-n":
-			startPromptIndex++
-			Flags.noexec = true
-		default:
-			// skip searching for switches if invalid arg is found (assume it is prompt)
-			validArg = false
-			break
-		}
-	}
-
-	// noexec takes priority over autoexecute, turn off autoexec
-	// guarentees mutual exclusivity
-	if Flags.noexec && Flags.autoexecute {
-		Flags.autoexecute = false
-	}
-
-	prompt = strings.Join(args[startPromptIndex:], " ")
-	if prompt == "" {
-		return "", errors.New("Prompt cannot be empty") 
-	}
-
-	return prompt, nil
-}
 
 /**********************
  * main
@@ -125,28 +25,28 @@ var argParse = func (args []string, Flags *FlagStruct) (prompt string, err error
 
 func main() {
 
-	Flags := new_flags()
+	Flags := helpers.ConstructFlags()
 
 	// parse arguments and recieve prompt
-	prompt, err := argParse(os.Args, &Flags)
+	prompt, err := helpers.ArgParse(os.Args, &Flags)
 
 	// exit if -h/--help flags found
-	if Flags.help == true {
+	if Flags.Help == true {
 		// handler for when --help or -h are provided
-		Usage()
+		helpers.Usage()
 		os.Exit(0)
 	}
 
 	if err != nil {
 		fmt.Printf("Parse error: %s\n", err)
-		Usage()
+		helpers.Usage()
 		os.Exit(1)
 	}
 
-	print(Flags.verbose, "Prompt extracted:", prompt)
+	helpers.Print(Flags.Verbose, "Prompt extracted:", prompt)
 
 	// Make the API call
-	print(Flags.verbose, "Sending prompt...")
+	helpers.Print(Flags.Verbose, "Sending prompt...")
 
 	response, err := http.Post(apiUrl, "application/json", strings.NewReader(fmt.Sprintf(`{"prompt": "%s"}`, prompt)))
 	if err != nil {
@@ -156,7 +56,7 @@ func main() {
 	defer response.Body.Close()
 
 	// Parse the response body as JSON
-	print(Flags.verbose, "Parsing response...")
+	helpers.Print(Flags.Verbose, "Parsing response...")
 
 	var data map[string]interface{}
 	err = json.NewDecoder(response.Body).Decode(&data)
@@ -166,7 +66,7 @@ func main() {
 	}
 
 	// Check if the "output" field exists
-	print(Flags.verbose, "Checking AI output field...")
+	helpers.Print(Flags.Verbose, "Checking AI output field...")
 	
 	result, ok := data["output"].(string)
 	if !ok {
@@ -175,14 +75,14 @@ func main() {
 	}
 	
 	// Emit the result
-	print(Flags.verbose, "Output: \n")
+	helpers.Print(Flags.Verbose, "Output: \n")
 
 	fmt.Println(result)
 	fmt.Println()
 
 	// if not skipping prompt, ask user if they would like to execute
 	userExecute := false
-	if !Flags.noexec && !Flags.autoexecute {
+	if !Flags.Noexec && !Flags.Autoexecute {
 		var userInput string
 		fmt.Print("Would you like to execute the command? (y/n): ")
 		fmt.Scanln(&userInput)
@@ -193,7 +93,7 @@ func main() {
 	}
 
 	// perform the command if autoexecute enabled or user prompted to exec
-	if Flags.autoexecute || userExecute {
+	if Flags.Autoexecute || userExecute {
 		// convert to arr of values (exec requires a specific format)
 		fullCmd := strings.Split(result, " ")
 		cmd := fullCmd[0]
@@ -203,7 +103,7 @@ func main() {
 			args=fullCmd[1:]
 		}
 
-		print(Flags.verbose, "Executing the result...")
+		helpers.Print(Flags.Verbose, "Executing the result...")
 
 		out, err := exec.Command(cmd, args...).Output()
