+++ src/go.mod
@@ -1,0 +1,3 @@
+module fl
+
+go 1.22.5
